Rename: game package to controller
